Add files via upload
version 3: adds 4 new functions. append, first, last, and size. append adds a element to the back of a list and also increases the size by 50% if  the list is full. first returns the first element. last returns the last element. size returns the total size(length in this case) of the list